requirements.txt
streamlit
pandas
python-pptx
sentence-transformers
nltk
pdfminer.six
PyPDF2
pdf2image
Pillow
openai
